upgrade boot version

+++ src/main/resources/db/migration/mysql/V1.0.0__create_todos_table.sql
@@ -1,6 +1,6 @@
 DROP TABLE IF EXISTS `todos`;
 CREATE TABLE `todos` (
-  `id`        INT AUTO_INCREMENT,
+  `id`        BIGINT AUTO_INCREMENT,
   `title`     VARCHAR(255) DEFAULT NULL,
   `completed` BIT DEFAULT FALSE,
   PRIMARY KEY (`id`),
